backend/pyproject.toml
[project]
name = "mirofish-offline-backend"
version = "0.2.0"
description = "MiroFish-Offline - Offline-first fork running on Neo4j + Ollama"
requires-python = ">=3.11"
license = { text = "AGPL-3.0" }
authors = [
    { name = "MiroFish Team" }
]

dependencies = [
    # Core framework
    "flask>=3.0.0",
    "flask-cors>=6.0.0",

    # LLM
    "openai>=1.0.0",

    # Neo4j graph database driver
    "neo4j>=5.15.0",

    # OASIS social media simulation
    "camel-oasis==0.2.5",
    "camel-ai==0.2.78",

    # File processing
    "PyMuPDF>=1.24.0",
    # Encoding detection (non-UTF-8 text files)
    "charset-normalizer>=3.0.0",
    "chardet>=5.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",

    # SQLite storage (SQLAlchemy Core)
    "sqlalchemy>=2.0.0",

    # Calibration (Platt scaling)
    "scikit-learn>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pipreqs>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
